% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSummarizedExperiment.R
\name{loadSummarizedExperiment}
\alias{loadSummarizedExperiment}
\title{Load a SummarizedExperiment}
\usage{
loadSummarizedExperiment(exp.info, project)
}
\arguments{
\item{exp.info}{Named list containing the metadata for this experiment.}

\item{project}{Any argument accepted by the acquisition functions, see \code{?\link{acquireFile}}. 
By default, this should be a string containing the path to a staging directory.}
}
\value{
A \linkS4class{SummarizedExperiment} or \linkS4class{RangedSummarizedExperiment} object.
}
\description{
Default loading of \linkS4class{SummarizedExperiment}s based on the metadata stored by the corresponding \code{\link{stageObject}} method.
}
\examples{
# Mocking up an experiment:
mat <- matrix(rpois(10000, 10), ncol=10)
colnames(mat) <- letters[1:10]
rownames(mat) <- sprintf("GENE_\%i", seq_len(nrow(mat)))

se <- SummarizedExperiment(list(counts=mat))
se$stuff <- LETTERS[1:10]
rowData(se)$blah <- runif(1000)
metadata(se)$whee <- "YAY"

# Staging it:
tmp <- tempfile()
dir.create(tmp)
info <- stageObject(se, dir=tmp, "rna-seq") 

# And loading it back in:
loadSummarizedExperiment(info, tmp)

}
\author{
Aaron Lun
}
