% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSummarizedExperiment.R
\name{readSummarizedExperiment}
\alias{readSummarizedExperiment}
\alias{loadSummarizedExperiment}
\title{Read a SummarizedExperiment from disk}
\usage{
readSummarizedExperiment(path, metadata, ...)
}
\arguments{
\item{path}{String containing a path to a directory, itself created using the \code{\link[alabaster.base]{saveObject}} method for SummarizedExperiment objects.}

\item{metadata}{Named list of metadata for this object, see \code{\link[alabaster.base]{readObjectFile}} for details.}

\item{...}{Further arguments passed to internal \code{\link[alabaster.base]{altReadObject}} calls.}
}
\value{
A SummarizedExperiment object.
}
\description{
Read a \link[SummarizedExperiment]{SummarizedExperiment} from its on-disk representation.
This is usually not directly called by users, but is instead called by dispatch in \code{\link[alabaster.base]{readObject}}.
}
\examples{
# Mocking up an experiment:
mat <- matrix(rpois(10000, 10), ncol=10)
colnames(mat) <- letters[1:10]
rownames(mat) <- sprintf("GENE_\%i", seq_len(nrow(mat)))

se <- SummarizedExperiment(list(counts=mat))
se$stuff <- LETTERS[1:10]
rowData(se)$blah <- runif(1000)
metadata(se)$whee <- "YAY"

tmp <- tempfile()
saveObject(se, tmp)
readObject(tmp)

}
\seealso{
\code{"\link{saveObject,SummarizedExperiment-method}"}, to save the SummarizedExperiment to disk.
}
\author{
Aaron Lun
}
