% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stageSummarizedExperiment.R
\name{stageObject,SummarizedExperiment-method}
\alias{stageObject,SummarizedExperiment-method}
\alias{stageObject,RangedSummarizedExperiment-method}
\title{Stage a SummarizedExperiment}
\usage{
\S4method{stageObject}{SummarizedExperiment}(x, dir, path, child = FALSE, meta.name = "experiment.json", ...)

\S4method{stageObject}{RangedSummarizedExperiment}(x, dir, path, child = FALSE, ..., skip.ranges = FALSE)
}
\arguments{
\item{x}{A \linkS4class{SummarizedExperiment} object or one of its subclasses.}

\item{dir}{String containing the path to the staging directory.}

\item{path}{String containing a prefix of the relative path inside \code{dir} where \code{x} is to be saved.
The actual path used to save \code{x} may include additional components, see Details.}

\item{child}{Logical scalar indicating whether \code{x} is a child of a larger object.}

\item{meta.name}{String containing the name of the metadata file.}

\item{...}{Further arguments to pass to the \linkS4class{SummarizedExperiment} method.
For the SummarizedExperiment itself, all further arguments are just ignored.}

\item{skip.ranges}{Logical scalar indicating whether to avoid saving the \code{\link{rowRanges}}.}
}
\value{
A named list of metadata that follows the \code{summarized_experiment} schema.
The contents of \code{x} are saved into a \code{path} subdirectory inside \code{dir}.
}
\description{
Save a \linkS4class{SummarizedExperiment} to file inside the staging directory.
}
\details{
\code{meta.name} is only needed to set up the output \code{path}, for consistency with the \code{\link{stageObject}} contract.
Callers should make sure to write the metadata to the same path by using \code{\link{.writeMetadata}} to create the JSON file.

If \code{skip.ranges=TRUE}, the RangedSummarizedExperiment method just calls the SummarizedExperiment method, i.e., \code{\link{rowRanges}} are not saved.
This avoids the hassle of switching classes and the associated problems, e.g., \url{https://github.com/Bioconductor/SummarizedExperiment/issues/29}.

By default, we consider the presence of data frames in the assays to be an error.
Users should coerce these into an appropriate matrix type, e.g., a dense matrix or a sparse dgCMatrix.
If a DataFrame as an assay is truly desired, users may set \code{\link{options}(alabaster.se.reject_data.frames=FALSE)} to skip the error.
Note that this only works for \linkS4class{DataFrame} objects - data.frame objects will not be saved correctly.
}
\examples{
tmp <- tempfile()
dir.create(tmp)

mat <- matrix(rpois(10000, 10), ncol=10)
colnames(mat) <- letters[1:10]
rownames(mat) <- sprintf("GENE_\%i", seq_len(nrow(mat)))

se <- SummarizedExperiment(list(counts=mat))
se$stuff <- LETTERS[1:10]
rowData(se)$blah <- runif(1000)
metadata(se)$whee <- "YAY"

dir.create(tmp)
stageObject(se, dir=tmp, "rna-seq") 
list.files(file.path(tmp, "rna-seq"))

}
\author{
Aaron Lun
}
